add trivial way to handle accepted connections
Since accepted connections need to be manually closed on exit
for the other end to detect the node as disconnected, a very simple
way to handle accepted clients was implemented for that purpose only.
This should not remain in the future since it creates and maintain 2
TCP connections just to connect 2 peers together (when 1 should be enough).

    - node.h: add a_bevs and a_size fields to node_comm struct
    - node.c: init and free those 2 new fields
              pass the node struct as argument to the listener's callback
    - events.c: implement the accept_conn_cb of the listener
                    realloc a_bevs when a new client connects
                    create a new bev from the new sock for that client